Move `groupBy` out of `search` and into `functions`
